[CL-1941] Move to admin
