+++ front/app/components/admin/ContentBuilder/Editor/index.tsx
