+++ front/app/components/admin/ContentBuilder/Frame/ContentBuilderFrame.test.tsx
