+++ front/app/components/admin/ContentBuilder/Toolbox/DraggableElement.tsx
