+++ front/app/components/admin/ContentBuilder/Toolbox/SectionTitle.tsx
