+++ front/app/components/admin/ContentBuilder/TopBar/GoBackButton.tsx
