+++ front/app/components/admin/ContentBuilder/Widgets/AboutBox/index.tsx
