+++ front/app/components/admin/ContentBuilder/Widgets/Accordion/index.tsx
