+++ front/app/components/admin/ContentBuilder/Widgets/Accordion/messages.ts
