+++ front/app/components/admin/ContentBuilder/Widgets/Image/messages.ts
