+++ front/app/components/admin/ContentBuilder/Widgets/Text/index.tsx
