+++ front/app/components/admin/ContentBuilder/Widgets/TwoColumn/index.tsx
