+++ front/app/components/admin/ContentBuilder/Widgets/TwoColumn/messages.ts
