+++ front/app/components/admin/ContentBuilder/Widgets/WhiteSpace/index.tsx
