+++ front/app/modules/commercial/content_builder/admin/components/ContentViewer/Viewer/index.tsx
@@ -7,7 +7,7 @@ import useLocalize from 'hooks/useLocalize';
 
 // components
 import Editor from '../../Editor';
-import ContentBuilderFrame from 'components/ContentBuilder/Frame';
+import ContentBuilderFrame from 'components/admin/ContentBuilder/Frame';
 import { Box, Spinner, Title } from '@citizenlab/cl2-component-library';
 import ProjectInfo from 'containers/ProjectsShowPage/shared/header/ProjectInfo';
 import { isNilOrError } from 'utils/helperUtils';
