+++ front/app/modules/commercial/content_builder/admin/containers/FullscreenPreview/index.tsx
@@ -10,7 +10,7 @@ import { useParams } from 'react-router-dom';
 
 // components
 import Editor from '../../components/Editor';
-import ContentBuilderFrame from 'components/ContentBuilder/Frame';
+import ContentBuilderFrame from 'components/admin/ContentBuilder/Frame';
 import { Box, Spinner, Title } from '@citizenlab/cl2-component-library';
 import { isNilOrError } from 'utils/helperUtils';
